R6.2: split vector/clock_ad into model/build/check/simulation
Facade mod.rs re-exports everything from the former monolith; no public
path changes. Producer derivations (islands, transports, reverse-AD
fallbacks) live in build.rs; the standalone checker and the shared
terminal verify path live in check.rs; reference simulators in
simulation.rs; vocabulary + error taxonomy in model.rs.

Plan §4.8 — admission guards in the shared verify function
`verify_vector_clock_ad_plan_after_vector_plan` (hosted in check.rs):
- reject_unadopted_stateful_reads

remains on BOTH paths after the split: the producer's terminal
verification (build_vector_clock_ad_plan) and the standalone checker
(verify_vector_clock_ad_plan). New rejection tests cover both entries,
including one that fails through the checker entry point alone
(unadopted_stateful_read_rejected_through_checker_entry_alone).

Validation: transform 387 lib tests, vector_mode oracle 35/35,
workspace clippy (+1.97.0) clean.
