+++ crates/transform/src/signal_fir/vector/clock_ad/mod.rs
@@ -24,1013 +24,15 @@
 //! does not claim vector reverse-window semantics and cannot activate a
 //! backend path by itself.
 
-use std::collections::{BTreeMap, BTreeSet};
-use std::fmt;
+pub mod build;
+pub mod check;
+pub mod model;
+pub mod simulation;
 
-use propagate::{ClockDomainKind, ClockDomainTable};
-use signals::{SigId, SigMatch, match_sig};
-use sigtype::Nature;
-use tlib::{match_sym_rec, match_sym_ref};
-
-use crate::clk_env::{ClkEnvError, annotate};
-use crate::signal_prepare::VerifiedPreparedSignals;
-
-use super::analysis::{DepKind, EffectAtom, StateCell, StateResource};
-use super::decoration_verify::{CanonicalSigType, VerifiedDecorationCertificate};
-use super::plan::VerifiedVectorPlan;
-use super::verify::{LoopKind, Placement, VectorPlan, VectorPlanError, verify_vector_plan};
-
-/// Current canonical P6.2 clock/AD-plan schema.
-pub const VECTOR_CLOCK_AD_PLAN_VERSION: u32 = 1;
-
-/// Concrete guard shape used to implement `fires(c, i)`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum ClockGuard {
-    BooleanOnDemand,
-    CountedOnDemand,
-    CountedUpsampling,
-    DownsampleModulo,
-}
-
-/// Whether a pre-planned P5 transport may use the outer chunk index.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum ClockTransportMode {
-    /// Audio-rate value: the P5 `i0 - vindex` chunk route remains valid.
-    OuterChunk,
-    /// Per-sample value crossing logical loops in one certified fused group.
-    /// The route is a scalar stack temporary inside the group's sole sample
-    /// envelope, either a top-rate loop or one exact guarded clock island,
-    /// never a chunk array filled by a preceding loop.
-    FusedScalar { group_id: u64 },
-    /// Domain-rate value: route below the serial guard, one fire at a time.
-    IslandScalar { domain_id: u64 },
-    /// Persistent `PermVar` value exported from a domain to an ancestor.
-    HeldOutput { domain_id: u64 },
-}
-
-/// Complete policy for one existing P5 transport.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct ClockTransportPolicy {
-    pub transport_id: u64,
-    pub mode: ClockTransportMode,
-}
-
-/// One serial OD/US/DS region and the P4 loops nested below it.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ClockIsland {
-    pub domain_id: u64,
-    pub parent_domain: Option<u64>,
-    pub kind: ClockDomainKind,
-    pub clock_signal_id: u64,
-    pub wrapper_signal_id: u64,
-    pub boundary_loop_id: u64,
-    pub guard: ClockGuard,
-    /// Exact signals whose inferred clock environment is this domain.
-    pub signal_ids: Vec<u64>,
-    /// Signals that own clock guard/counter state for this domain.
-    pub clock_state_signal_ids: Vec<u64>,
-    /// Existing P4 loops with at least one root in this domain.
-    pub nested_loop_ids: Vec<u64>,
-}
-
-/// FAD policy at the prepared-signal boundary.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ForwardAdPolicy {
-    /// Propagation expanded FAD into ordinary primal/tangent signals.
-    ExpandedSignalGraph,
-}
-
-/// Reverse-mode carrier requiring a block-local reverse window.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum ReverseAdKind {
-    ReverseTimeRec,
-    BlockReverseAd,
-}
-
-/// Semantically fixed reverse-mode epoch.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum AdEpoch {
-    Forward,
-    Reverse,
-}
-
-/// Stable reason why reverse-mode execution is not admitted to vector mode.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ReverseAdDiagnostic {
-    ScalarReverseWindowRequired,
-}
-
-impl ReverseAdDiagnostic {
-    /// Stable internal diagnostic code for a future CLI/backend activation gate.
-    #[must_use]
-    pub fn code(self) -> &'static str {
-        match self {
-            Self::ScalarReverseWindowRequired => "FRS-VEC-RAD-SCALAR",
-        }
-    }
-
-    /// User-facing reason that vector reverse-mode execution is refused.
-    #[must_use]
-    pub fn message(self) -> &'static str {
-        match self {
-            Self::ScalarReverseWindowRequired => {
-                "reverse AD requires a scalar forward/tape/reverse window; vector chunk semantics are not enabled"
-            }
-        }
-    }
-}
-
-/// One explicit scalar fallback for a reverse-time Signal-IR carrier.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ReverseAdFallback {
-    pub signal_id: u64,
-    pub owner_loop_id: u64,
-    pub kind: ReverseAdKind,
-    pub epochs: Vec<AdEpoch>,
-    pub diagnostic: ReverseAdDiagnostic,
-}
-
-/// Canonical finite P6.2 artifact.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct VectorClockAdPlan {
-    pub schema_version: u32,
-    pub vec_size: u64,
-    pub clock_islands: Vec<ClockIsland>,
-    pub transports: Vec<ClockTransportPolicy>,
-    pub forward_ad: ForwardAdPolicy,
-    pub reverse_ad_fallbacks: Vec<ReverseAdFallback>,
-}
-
-/// Opaque evidence that P6.2 construction passed its checker.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct VerifiedVectorClockAdPlan {
-    plan: VectorClockAdPlan,
-    vector_plan: VectorPlan,
-}
-
-impl VerifiedVectorClockAdPlan {
-    #[must_use]
-    pub fn plan(&self) -> &VectorClockAdPlan {
-        &self.plan
-    }
-
-    #[must_use]
-    pub fn vector_plan(&self) -> &VectorPlan {
-        &self.vector_plan
-    }
-
-    #[must_use]
-    pub fn into_plan(self) -> VectorClockAdPlan {
-        self.plan
-    }
-
-    /// State resources whose execution is completely owned by P6.2 guards or
-    /// persistent held-output transports.
-    #[must_use]
-    pub fn managed_state_resources(&self) -> BTreeSet<StateResource> {
-        self.vector_plan
-            .signals
-            .iter()
-            .flat_map(|signal| signal.effects.iter())
-            .filter_map(|effect| match effect {
-                EffectAtom::ReadState(resource) | EffectAtom::WriteState(resource)
-                    if matches!(
-                        resource,
-                        StateResource::Signal {
-                            cell: StateCell::Clock | StateCell::Hold,
-                            ..
-                        }
-                    ) =>
-                {
-                    Some(resource.clone())
-                }
-                _ => None,
-            })
-            .collect()
-    }
-}
-
-#[cfg(test)]
-pub(crate) fn verified_vector_clock_ad_plan_for_test(
-    plan: VectorClockAdPlan,
-    vector_plan: &VerifiedVectorPlan,
-) -> VerifiedVectorClockAdPlan {
-    assert_eq!(plan.schema_version, VECTOR_CLOCK_AD_PLAN_VERSION);
-    assert_eq!(plan.vec_size, vector_plan.plan().vec_size);
-    assert_eq!(plan.transports.len(), vector_plan.plan().transports.len());
-    for (policy, transport) in plan.transports.iter().zip(&vector_plan.plan().transports) {
-        assert_eq!(policy.transport_id, transport.transport_id);
-    }
-    VerifiedVectorClockAdPlan {
-        plan,
-        vector_plan: vector_plan.plan().clone(),
-    }
-}
-
-/// Typed producer/checker failure at the P6.2 boundary.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum VectorClockAdError {
-    Plan(VectorPlanError),
-    ClockInference(ClkEnvError),
-    UnsupportedSchema {
-        found: u32,
-    },
-    VecSizeMismatch {
-        declared: u64,
-        actual: u64,
-    },
-    SignalCoverageMismatch {
-        signal_id: u64,
-    },
-    ClockFactMismatch {
-        signal_id: u64,
-    },
-    DomainParentUnknown {
-        domain_id: u64,
-        parent_id: u64,
-    },
-    DomainCycle {
-        domain_id: u64,
-    },
-    ClockSignalUnknown {
-        domain_id: u64,
-        signal_id: u64,
-    },
-    UnsupportedClockType {
-        domain_id: u64,
-        kind: ClockDomainKind,
-    },
-    WrapperCoverageMismatch {
-        domain_id: u64,
-    },
-    WrapperKindMismatch {
-        domain_id: u64,
-        table: ClockDomainKind,
-        signal: ClockDomainKind,
-    },
-    ClockStateDomainUnknown {
-        signal_id: u64,
-    },
-    BoundaryNotOwned {
-        domain_id: u64,
-        signal_id: u64,
-    },
-    BoundaryNotSerial {
-        domain_id: u64,
-        loop_id: u64,
-    },
-    IslandCoverageMismatch,
-    TransportCoverageMismatch,
-    ReverseCarrierNotOwned {
-        signal_id: u64,
-    },
-    ReverseAdCoverageMismatch,
-    ClockValueKindMismatch {
-        guard: ClockGuard,
-    },
-    InvalidDownsampleFactor {
-        factor: i64,
-    },
-    UnadoptedStatefulRead {
-        domain_id: u64,
-        consumer: u64,
-        producer: u64,
-    },
-}
-
-impl fmt::Display for VectorClockAdError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Plan(error) => write!(f, "vector plan verification failed: {error}"),
-            Self::ClockInference(error) => write!(f, "clock inference failed: {error}"),
-            Self::UnsupportedSchema { found } => {
-                write!(f, "unsupported vector clock/AD schema {found}")
-            }
-            Self::VecSizeMismatch { declared, actual } => write!(
-                f,
-                "clock/AD plan vector size {declared} differs from vector plan {actual}"
-            ),
-            Self::SignalCoverageMismatch { signal_id } => {
-                write!(f, "clock/AD source facts do not cover signal {signal_id}")
-            }
-            Self::ClockFactMismatch { signal_id } => {
-                write!(f, "clock facts disagree for signal {signal_id}")
-            }
-            Self::DomainParentUnknown {
-                domain_id,
-                parent_id,
-            } => write!(
-                f,
-                "clock domain {domain_id} names unknown parent {parent_id}"
-            ),
-            Self::DomainCycle { domain_id } => {
-                write!(
-                    f,
-                    "clock domain hierarchy cycles through domain {domain_id}"
-                )
-            }
-            Self::ClockSignalUnknown {
-                domain_id,
-                signal_id,
-            } => write!(
-                f,
-                "clock domain {domain_id} names unknown clock signal {signal_id}"
-            ),
-            Self::UnsupportedClockType { domain_id, kind } => write!(
-                f,
-                "clock domain {domain_id} ({kind:?}) has no supported boolean/integer guard"
-            ),
-            Self::WrapperCoverageMismatch { domain_id } => write!(
-                f,
-                "clock domain {domain_id} does not have exactly one matching wrapper"
-            ),
-            Self::WrapperKindMismatch {
-                domain_id,
-                table,
-                signal,
-            } => write!(
-                f,
-                "clock domain {domain_id} is {table:?} in the table but {signal:?} in Signal IR"
-            ),
-            Self::ClockStateDomainUnknown { signal_id } => write!(
-                f,
-                "clock-state signal {signal_id} has no decodable domain token"
-            ),
-            Self::BoundaryNotOwned {
-                domain_id,
-                signal_id,
-            } => write!(
-                f,
-                "clock boundary {domain_id} signal {signal_id} has no owned loop"
-            ),
-            Self::BoundaryNotSerial { domain_id, loop_id } => write!(
-                f,
-                "clock boundary {domain_id} is owned by non-serial loop {loop_id}"
-            ),
-            Self::IslandCoverageMismatch => write!(f, "clock-island facts are not exact"),
-            Self::TransportCoverageMismatch => {
-                write!(f, "clock transport policies are not exact")
-            }
-            Self::ReverseCarrierNotOwned { signal_id } => write!(
-                f,
-                "reverse-mode carrier {signal_id} has no owned scalar loop"
-            ),
-            Self::ReverseAdCoverageMismatch => {
-                write!(f, "reverse-mode fallback facts are not exact")
-            }
-            Self::ClockValueKindMismatch { guard } => {
-                write!(f, "runtime clock value does not match {guard:?}")
-            }
-            Self::InvalidDownsampleFactor { factor } => {
-                write!(f, "downsampling factor must be positive, got {factor}")
-            }
-            Self::UnadoptedStatefulRead {
-                domain_id,
-                consumer,
-                producer,
-            } => {
-                write!(
-                    f,
-                    "clock domain {domain_id}: signal {consumer} reads audio-rate stateful \
-                     signal {producer} from fire time; rate-polymorphic state under a clock \
-                     wrapper is not adopted into the domain yet and would advance at the \
-                     wrong rate"
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for VectorClockAdError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            Self::Plan(error) => Some(error),
-            Self::ClockInference(error) => Some(error),
-            _ => None,
-        }
-    }
-}
-
-impl From<VectorPlanError> for VectorClockAdError {
-    fn from(value: VectorPlanError) -> Self {
-        Self::Plan(value)
-    }
-}
-
-impl From<ClkEnvError> for VectorClockAdError {
-    fn from(value: ClkEnvError) -> Self {
-        Self::ClockInference(value)
-    }
-}
-
-/// Builds and independently checks the P6.2 island/AD policy.
-pub fn build_vector_clock_ad_plan(
-    prepared: &VerifiedPreparedSignals,
-    domains: &ClockDomainTable,
-    decorations: &VerifiedDecorationCertificate,
-    vector_plan: &VerifiedVectorPlan,
-) -> Result<VerifiedVectorClockAdPlan, VectorClockAdError> {
-    let plan = vector_plan.plan();
-    verify_source_alignment(prepared, domains, decorations, plan)?;
-    let clock_islands = derive_clock_islands(prepared, domains, decorations, plan)?;
-    let transports = derive_transport_policies(prepared, plan)?;
-    let reverse_ad_fallbacks = derive_reverse_fallbacks(prepared, decorations, plan)?;
-    let clock_ad_plan = VectorClockAdPlan {
-        schema_version: VECTOR_CLOCK_AD_PLAN_VERSION,
-        vec_size: plan.vec_size,
-        clock_islands,
-        transports,
-        forward_ad: ForwardAdPolicy::ExpandedSignalGraph,
-        reverse_ad_fallbacks,
-    };
-    verify_vector_clock_ad_plan_after_vector_plan(
-        prepared,
-        domains,
-        decorations,
-        plan,
-        &clock_ad_plan,
-    )?;
-    Ok(VerifiedVectorClockAdPlan {
-        plan: clock_ad_plan,
-        vector_plan: plan.clone(),
-    })
-}
-
-/// Recomputes all island, transport, and reverse-fallback obligations.
-pub fn verify_vector_clock_ad_plan(
-    prepared: &VerifiedPreparedSignals,
-    domains: &ClockDomainTable,
-    decorations: &VerifiedDecorationCertificate,
-    vector_plan: &VectorPlan,
-    clock_ad_plan: &VectorClockAdPlan,
-) -> Result<(), VectorClockAdError> {
-    verify_vector_plan(vector_plan)?;
-    verify_vector_clock_ad_plan_after_vector_plan(
-        prepared,
-        domains,
-        decorations,
-        vector_plan,
-        clock_ad_plan,
-    )
-}
-
-fn verify_vector_clock_ad_plan_after_vector_plan(
-    prepared: &VerifiedPreparedSignals,
-    domains: &ClockDomainTable,
-    decorations: &VerifiedDecorationCertificate,
-    vector_plan: &VectorPlan,
-    clock_ad_plan: &VectorClockAdPlan,
-) -> Result<(), VectorClockAdError> {
-    if clock_ad_plan.schema_version != VECTOR_CLOCK_AD_PLAN_VERSION {
-        return Err(VectorClockAdError::UnsupportedSchema {
-            found: clock_ad_plan.schema_version,
-        });
-    }
-    if clock_ad_plan.vec_size != vector_plan.vec_size {
-        return Err(VectorClockAdError::VecSizeMismatch {
-            declared: clock_ad_plan.vec_size,
-            actual: vector_plan.vec_size,
-        });
-    }
-    verify_source_alignment(prepared, domains, decorations, vector_plan)?;
-    let expected_islands = derive_clock_islands(prepared, domains, decorations, vector_plan)?;
-    if clock_ad_plan.clock_islands != expected_islands {
-        return Err(VectorClockAdError::IslandCoverageMismatch);
-    }
-    reject_unadopted_stateful_reads(prepared, decorations, vector_plan, &expected_islands)?;
-    let expected_transports = derive_transport_policies(prepared, vector_plan)?;
-    if clock_ad_plan.transports != expected_transports {
-        return Err(VectorClockAdError::TransportCoverageMismatch);
-    }
-    if clock_ad_plan.forward_ad != ForwardAdPolicy::ExpandedSignalGraph {
-        return Err(VectorClockAdError::ReverseAdCoverageMismatch);
-    }
-    let expected_reverse = derive_reverse_fallbacks(prepared, decorations, vector_plan)?;
-    if clock_ad_plan.reverse_ad_fallbacks != expected_reverse {
-        return Err(VectorClockAdError::ReverseAdCoverageMismatch);
-    }
-    Ok(())
-}
-
-/// Rejects fire-time reads of audio-rate stateful values.
-///
-/// Clock inference is rate-polymorphic: a recursion or delay whose inputs are
-/// all constants infers the bottom environment, so the plan hoists it into an
-/// outer-rate loop even when it is written under a clock wrapper (e.g.
-/// `upsampling(1 : (+ ~ *(0.5)))`). Its state then advances once per outer
-/// sample instead of once per fire — a miscompile the reference semantics
-/// reject. Until such groups are adopted into their wrapper's domain, any
-/// value dependency from an in-domain consumer to an audio-rate stateful
-/// producer fails closed. The one legal crossing is the `ZeroPad` value
-/// operand: propagation inserts it exactly where an outer-rate producer is
-/// zero-stuffed into an upsampled domain.
-fn reject_unadopted_stateful_reads(
-    prepared: &VerifiedPreparedSignals,
-    decorations: &VerifiedDecorationCertificate,
-    plan: &VectorPlan,
-    islands: &[ClockIsland],
-) -> Result<(), VectorClockAdError> {
-    let ids = super::common::ids::prepared_signal_ids(prepared);
-    let signals = plan
-        .signals
-        .iter()
-        .map(|signal| (signal.signal_id, signal))
-        .collect::<BTreeMap<_, _>>();
-    let domain_by_clock = islands
-        .iter()
-        .map(|island| (island.domain_id + 1, island.domain_id))
-        .collect::<BTreeMap<_, _>>();
-    for dependency in &decorations.certificate().dependencies {
-        if !matches!(
-            dependency.kind,
-            DepKind::Immediate | DepKind::Delayed { .. }
-        ) {
-            continue;
-        }
-        let Some(consumer) = signals.get(&u64::from(dependency.from)) else {
-            continue;
-        };
-        let Some(producer) = signals.get(&u64::from(dependency.to)) else {
-            continue;
-        };
-        if producer.clock_id != 0 {
-            continue;
-        }
-        let Some(&domain_id) = domain_by_clock.get(&consumer.clock_id) else {
-            continue;
-        };
-        let Some(&producer_sig) = ids.get(&dependency.to) else {
-            continue;
-        };
-        if !is_stateful_producer(prepared, producer_sig) {
-            continue;
-        }
-        let consumer_sig = ids.get(&dependency.from).copied();
-        if consumer_sig
-            .is_some_and(|sig| matches!(match_sig(prepared.arena(), sig), SigMatch::ZeroPad(_, _)))
-        {
-            continue;
-        }
-        return Err(VectorClockAdError::UnadoptedStatefulRead {
-            domain_id,
-            consumer: u64::from(dependency.from),
-            producer: u64::from(dependency.to),
-        });
-    }
-    Ok(())
-}
-
-/// Whether the signal carries per-sample state whose advance rate is
-/// observable: symbolic recursion (group, reference, or projection) and
-/// explicit delays.
-fn is_stateful_producer(prepared: &VerifiedPreparedSignals, sig: SigId) -> bool {
-    let arena = prepared.arena();
-    if match_sym_rec(arena, sig).is_some() || match_sym_ref(arena, sig).is_some() {
-        return true;
-    }
-    match match_sig(arena, sig) {
-        SigMatch::Delay1(_) | SigMatch::Delay(_, _) | SigMatch::Prefix(_, _) => true,
-        SigMatch::Proj(_, group) => {
-            match_sym_rec(arena, group).is_some() || match_sym_ref(arena, group).is_some()
-        }
-        _ => false,
-    }
-}
-
-fn verify_source_alignment(
-    prepared: &VerifiedPreparedSignals,
-    domains: &ClockDomainTable,
-    decorations: &VerifiedDecorationCertificate,
-    plan: &VectorPlan,
-) -> Result<(), VectorClockAdError> {
-    let records = &decorations.certificate().records;
-    if records.len() != plan.signals.len() {
-        return Err(VectorClockAdError::SignalCoverageMismatch {
-            signal_id: u64::MAX,
-        });
-    }
-    let clocks = annotate(prepared.arena(), domains, prepared.outputs())?;
-    let ids = super::common::ids::prepared_signal_ids(prepared);
-    for (record, signal) in records.iter().zip(&plan.signals) {
-        let signal_id = u64::from(record.signal_id);
-        if signal.signal_id != signal_id {
-            return Err(VectorClockAdError::SignalCoverageMismatch { signal_id });
-        }
-        let sig = ids
-            .get(&record.signal_id)
-            .copied()
-            .ok_or(VectorClockAdError::SignalCoverageMismatch { signal_id })?;
-        let inferred = clocks
-            .env(sig)
-            .ok_or(VectorClockAdError::ClockFactMismatch { signal_id })?
-            .map(|domain| domain.as_u32());
-        if record.clock_domain != inferred
-            || signal.clock_id != inferred.map_or(0, |domain| u64::from(domain) + 1)
-        {
-            return Err(VectorClockAdError::ClockFactMismatch { signal_id });
-        }
-    }
-    verify_domain_tree(domains)
-}
-
-fn verify_domain_tree(domains: &ClockDomainTable) -> Result<(), VectorClockAdError> {
-    for (domain_id, domain) in domains.iter() {
-        let raw = u64::from(domain_id.as_u32());
-        if let Some(parent) = domain.parent
-            && domains.get(parent).is_none()
-        {
-            return Err(VectorClockAdError::DomainParentUnknown {
-                domain_id: raw,
-                parent_id: u64::from(parent.as_u32()),
-            });
-        }
-        let mut seen = BTreeSet::new();
-        let mut cursor = Some(domain_id);
-        while let Some(current) = cursor {
-            if !seen.insert(current.as_u32()) {
-                return Err(VectorClockAdError::DomainCycle { domain_id: raw });
-            }
-            cursor = domains.get(current).and_then(|entry| entry.parent);
-        }
-    }
-    Ok(())
-}
-
-fn derive_clock_islands(
-    prepared: &VerifiedPreparedSignals,
-    domains: &ClockDomainTable,
-    decorations: &VerifiedDecorationCertificate,
-    plan: &VectorPlan,
-) -> Result<Vec<ClockIsland>, VectorClockAdError> {
-    let ids = super::common::ids::prepared_signal_ids(prepared);
-    let records = decorations
-        .certificate()
-        .records
-        .iter()
-        .map(|record| (record.signal_id, record))
-        .collect::<BTreeMap<_, _>>();
-    let signals = plan
-        .signals
-        .iter()
-        .map(|signal| (signal.signal_id, signal))
-        .collect::<BTreeMap<_, _>>();
-    let loops = plan
-        .loops
-        .iter()
-        .map(|record| (record.loop_id, record))
-        .collect::<BTreeMap<_, _>>();
-
-    let mut wrappers = BTreeMap::<u64, Vec<(u64, u32, ClockDomainKind)>>::new();
-    let mut clock_state = BTreeMap::<u64, Vec<u64>>::new();
-    for record in records.values() {
-        let sig = ids[&record.signal_id];
-        if let Some((domain, clock, kind)) = wrapper_domain_and_clock(prepared, sig) {
-            wrappers.entry(domain).or_default().push((
-                u64::from(record.signal_id),
-                clock.as_u32(),
-                kind,
-            ));
-        }
-        if record
-            .effects
-            .iter()
-            .any(|effect| is_owned_clock_effect(effect, record.signal_id))
-        {
-            if matches!(match_sig(prepared.arena(), sig), SigMatch::Clocked(token, _)
-                if prepared.arena().is_nil(token))
-            {
-                continue;
-            }
-            let domain = clock_state_domain(prepared, sig).ok_or(
-                VectorClockAdError::ClockStateDomainUnknown {
-                    signal_id: u64::from(record.signal_id),
-                },
-            )?;
-            clock_state
-                .entry(domain)
-                .or_default()
-                .push(u64::from(record.signal_id));
-        }
-    }
-
-    let mut islands = Vec::new();
-    for (domain_id, domain) in domains.iter() {
-        let raw = u64::from(domain_id.as_u32());
-        let domain_wrappers = wrappers.get(&raw).cloned().unwrap_or_default();
-        if domain_wrappers.len() != 1 {
-            return Err(VectorClockAdError::WrapperCoverageMismatch { domain_id: raw });
-        }
-        let (wrapper_signal_id, clock_signal, signal_kind) = domain_wrappers[0];
-        if signal_kind != domain.kind {
-            return Err(VectorClockAdError::WrapperKindMismatch {
-                domain_id: raw,
-                table: domain.kind,
-                signal: signal_kind,
-            });
-        }
-        let clock_signal_id = u64::from(clock_signal);
-        let clock_record =
-            records
-                .get(&clock_signal)
-                .ok_or(VectorClockAdError::ClockSignalUnknown {
-                    domain_id: raw,
-                    signal_id: clock_signal_id,
-                })?;
-        let guard = guard_for(raw, domain.kind, &clock_record.sig_type)?;
-        let wrapper = signals[&wrapper_signal_id];
-        let Placement::Owned(boundary_loop_id) = wrapper.placement else {
-            return Err(VectorClockAdError::BoundaryNotOwned {
-                domain_id: raw,
-                signal_id: wrapper_signal_id,
-            });
-        };
-        if !matches!(loops[&boundary_loop_id].kind, LoopKind::Island(_)) {
-            return Err(VectorClockAdError::BoundaryNotSerial {
-                domain_id: raw,
-                loop_id: boundary_loop_id,
-            });
-        }
-        let signal_ids = records
-            .values()
-            .filter(|record| record.clock_domain == Some(domain_id.as_u32()))
-            .map(|record| u64::from(record.signal_id))
-            .collect::<Vec<_>>();
-        let domain_clock_id = raw + 1;
-        let nested_loop_ids = plan
-            .loops
-            .iter()
-            .filter(|loop_record| {
-                loop_record
-                    .roots
-                    .iter()
-                    .any(|root| signals[root].clock_id == domain_clock_id)
-            })
-            .map(|loop_record| loop_record.loop_id)
-            .collect::<Vec<_>>();
-        islands.push(ClockIsland {
-            domain_id: raw,
-            parent_domain: domain.parent.map(|parent| u64::from(parent.as_u32())),
-            kind: domain.kind,
-            clock_signal_id,
-            wrapper_signal_id,
-            boundary_loop_id,
-            guard,
-            signal_ids,
-            clock_state_signal_ids: clock_state.remove(&raw).unwrap_or_default(),
-            nested_loop_ids,
-        });
-    }
-    if !clock_state.is_empty()
-        || wrappers
-            .keys()
-            .any(|domain| *domain >= domains.len() as u64)
-    {
-        return Err(VectorClockAdError::IslandCoverageMismatch);
-    }
-    Ok(islands)
-}
-
-fn derive_transport_policies(
-    prepared: &VerifiedPreparedSignals,
-    plan: &VectorPlan,
-) -> Result<Vec<ClockTransportPolicy>, VectorClockAdError> {
-    let ids = super::common::ids::prepared_signal_ids(prepared);
-    let signals = plan
-        .signals
-        .iter()
-        .map(|signal| (signal.signal_id, signal))
-        .collect::<BTreeMap<_, _>>();
-    let fused_transports = plan
-        .fused_serial_groups
-        .iter()
-        .flat_map(|group| {
-            group
-                .internal_transport_ids
-                .iter()
-                .map(move |&transport_id| (transport_id, group.group_id))
-        })
-        .collect::<BTreeMap<_, _>>();
-    plan.transports
-        .iter()
-        .map(|transport| {
-            let signal = signals.get(&transport.signal_id).ok_or(
-                VectorClockAdError::SignalCoverageMismatch {
-                    signal_id: transport.signal_id,
-                },
-            )?;
-            let mode = if let Some(&group_id) = fused_transports.get(&transport.transport_id) {
-                ClockTransportMode::FusedScalar { group_id }
-            } else if signal.clock_id == 0 {
-                ClockTransportMode::OuterChunk
-            } else if ids
-                .get(&u32::try_from(signal.signal_id).map_err(|_| {
-                    VectorClockAdError::SignalCoverageMismatch {
-                        signal_id: signal.signal_id,
-                    }
-                })?)
-                .is_some_and(|&sig| {
-                    matches!(match_sig(prepared.arena(), sig), SigMatch::PermVar(_))
-                })
-            {
-                ClockTransportMode::HeldOutput {
-                    domain_id: signal.clock_id - 1,
-                }
-            } else {
-                ClockTransportMode::IslandScalar {
-                    domain_id: signal.clock_id - 1,
-                }
-            };
-            Ok(ClockTransportPolicy {
-                transport_id: transport.transport_id,
-                mode,
-            })
-        })
-        .collect()
-}
-
-fn derive_reverse_fallbacks(
-    prepared: &VerifiedPreparedSignals,
-    decorations: &VerifiedDecorationCertificate,
-    plan: &VectorPlan,
-) -> Result<Vec<ReverseAdFallback>, VectorClockAdError> {
-    let ids = super::common::ids::prepared_signal_ids(prepared);
-    let signals = plan
-        .signals
-        .iter()
-        .map(|signal| (signal.signal_id, signal))
-        .collect::<BTreeMap<_, _>>();
-    let mut fallbacks = Vec::new();
-    for record in &decorations.certificate().records {
-        let sig = ids[&record.signal_id];
-        let kind = match match_sig(prepared.arena(), sig) {
-            SigMatch::ReverseTimeRec(_) => ReverseAdKind::ReverseTimeRec,
-            SigMatch::BlockReverseAD { .. } => ReverseAdKind::BlockReverseAd,
-            _ => continue,
-        };
-        let signal_id = u64::from(record.signal_id);
-        let Placement::Owned(owner_loop_id) = signals[&signal_id].placement else {
-            return Err(VectorClockAdError::ReverseCarrierNotOwned { signal_id });
-        };
-        fallbacks.push(ReverseAdFallback {
-            signal_id,
-            owner_loop_id,
-            kind,
-            epochs: vec![AdEpoch::Forward, AdEpoch::Reverse],
-            diagnostic: ReverseAdDiagnostic::ScalarReverseWindowRequired,
-        });
-    }
-    Ok(fallbacks)
-}
-
-fn wrapper_domain_and_clock(
-    prepared: &VerifiedPreparedSignals,
-    sig: SigId,
-) -> Option<(u64, SigId, ClockDomainKind)> {
-    let (children, kind) = match match_sig(prepared.arena(), sig) {
-        SigMatch::OnDemand(children) => (children, ClockDomainKind::OnDemand),
-        SigMatch::Upsampling(children) => (children, ClockDomainKind::Upsampling),
-        SigMatch::Downsampling(children) => (children, ClockDomainKind::Downsampling),
-        _ => return None,
-    };
-    let first = children.first().copied()?;
-    let SigMatch::Clocked(token, clock) = match_sig(prepared.arena(), first) else {
-        return None;
-    };
-    let SigMatch::ClockEnvToken(domain) = match_sig(prepared.arena(), token) else {
-        return None;
-    };
-    Some((u64::from(domain), clock, kind))
-}
-
-fn clock_state_domain(prepared: &VerifiedPreparedSignals, sig: SigId) -> Option<u64> {
-    match match_sig(prepared.arena(), sig) {
-        SigMatch::Clocked(token, _) => {
-            let SigMatch::ClockEnvToken(domain) = match_sig(prepared.arena(), token) else {
-                return None;
-            };
-            Some(u64::from(domain))
-        }
-        SigMatch::OnDemand(_) | SigMatch::Upsampling(_) | SigMatch::Downsampling(_) => {
-            wrapper_domain_and_clock(prepared, sig).map(|(domain, _, _)| domain)
-        }
-        _ => None,
-    }
-}
-
-fn is_owned_clock_effect(effect: &EffectAtom, signal_id: u32) -> bool {
-    matches!(
-        effect,
-        EffectAtom::ReadState(StateResource::Signal {
-            owner,
-            cell: StateCell::Clock,
-        }) | EffectAtom::WriteState(StateResource::Signal {
-            owner,
-            cell: StateCell::Clock,
-        })
-        if *owner == signal_id
-    )
-}
-
-fn guard_for(
-    domain_id: u64,
-    kind: ClockDomainKind,
-    ty: &CanonicalSigType,
-) -> Result<ClockGuard, VectorClockAdError> {
-    let (nature, boolean_interval) = scalar_clock_facts(ty)
-        .ok_or(VectorClockAdError::UnsupportedClockType { domain_id, kind })?;
-    match kind {
-        ClockDomainKind::OnDemand if boolean_interval => Ok(ClockGuard::BooleanOnDemand),
-        ClockDomainKind::OnDemand if nature == Nature::Int => Ok(ClockGuard::CountedOnDemand),
-        ClockDomainKind::Upsampling if nature == Nature::Int => Ok(ClockGuard::CountedUpsampling),
-        ClockDomainKind::Downsampling if nature == Nature::Int => Ok(ClockGuard::DownsampleModulo),
-        _ => Err(VectorClockAdError::UnsupportedClockType { domain_id, kind }),
-    }
-}
-
-fn scalar_clock_facts(ty: &CanonicalSigType) -> Option<(Nature, bool)> {
-    match ty {
-        CanonicalSigType::Sound => None,
-        CanonicalSigType::Simple {
-            nature, interval, ..
-        } => {
-            let lo = f64::from_bits(interval.lo_bits);
-            let hi = f64::from_bits(interval.hi_bits);
-            Some((
-                *nature,
-                !lo.is_nan() && !hi.is_nan() && lo >= 0.0 && hi <= 1.0,
-            ))
-        }
-        CanonicalSigType::Table { .. } | CanonicalSigType::Tuplet { .. } => None,
-    }
-}
-
-/// Runtime clock value used by the executable `ClockStep` reference model.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum ClockValue {
-    Boolean(bool),
-    Integer(i64),
-}
-
-/// Minimal concrete state needed to test fire-time and held-output semantics.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct ClockRuntime<S, O> {
-    pub state: S,
-    pub held_output: O,
-    pub downsample_counter: u64,
-}
-
-/// Applies `Step_c` exactly `fires(c,i)` times for one outer sample.
-///
-/// The held output is changed only by an inner transition. Consequently zero
-/// fires preserve both domain state and the previous output.
-pub fn simulate_clock_step<S, O, F>(
-    guard: ClockGuard,
-    clock: ClockValue,
-    runtime: &mut ClockRuntime<S, O>,
-    mut transition: F,
-) -> Result<u64, VectorClockAdError>
-where
-    F: FnMut(&S, u64) -> (S, O),
-{
-    let fires = match (guard, clock) {
-        (ClockGuard::BooleanOnDemand, ClockValue::Boolean(active)) => u64::from(active),
-        (
-            ClockGuard::CountedOnDemand | ClockGuard::CountedUpsampling,
-            ClockValue::Integer(count),
-        ) => u64::try_from(count).unwrap_or(0),
-        (ClockGuard::DownsampleModulo, ClockValue::Integer(factor)) => {
-            if factor <= 0 {
-                return Err(VectorClockAdError::InvalidDownsampleFactor { factor });
-            }
-            let fires = u64::from(runtime.downsample_counter == 0);
-            let factor = u64::try_from(factor).expect("positive i64 fits u64");
-            runtime.downsample_counter = (runtime.downsample_counter + 1) % factor;
-            fires
-        }
-        (guard, _) => return Err(VectorClockAdError::ClockValueKindMismatch { guard }),
-    };
-    for fire in 0..fires {
-        let (next_state, output) = transition(&runtime.state, fire);
-        runtime.state = next_state;
-        runtime.held_output = output;
-    }
-    Ok(fires)
-}
-
-/// Executes one scalar reverse window with immutable `Forward < Reverse` order.
-pub fn execute_reverse_ad_window<S, P, T, A, Forward, Reverse>(
-    initial_state: S,
-    forward: Forward,
-    reverse: Reverse,
-) -> (S, P, A)
-where
-    Forward: FnOnce(S) -> (S, P, T),
-    Reverse: FnOnce(S, T) -> (S, A),
-{
-    let (forward_state, primal, tape) = forward(initial_state);
-    let (reverse_state, adjoints) = reverse(forward_state, tape);
-    (reverse_state, primal, adjoints)
-}
+pub use build::*;
+pub use check::*;
+pub use model::*;
+pub use simulation::*;
 
 #[cfg(test)]
 mod tests;
