+++ porting/journal/2026-07-20.md
@@ -25,3 +25,27 @@ is complete: producer and checker reachability now live in visibly disjoint
 modules on both sides of the plan artifact. Validation: transform 385,
 `vector_mode` oracle 35/35, clippy `-D warnings` clean, fmt applied.
 
+
+## R6.2 — scission de `vector/clock_ad`
+
+- `clock_ad/mod.rs` (1 036 lignes) scindé en `model.rs` (374), `build.rs`
+  (351), `check.rs` (214), `simulation.rs` (69), façade `mod.rs` (38) —
+  méthode établie (parseur d'items, affectation explicite, visibilités
+  guidées par le compilateur, `#[cfg(test)]` sur l'import du constructeur
+  de test).
+- Règle §4.8 : `verify_vector_clock_ad_plan_after_vector_plan` (qui héberge
+  la garde d'admission `reject_unadopted_stateful_reads`) vit dans
+  `check.rs` et reste appelée par la vérification terminale du producteur
+  (`build.rs::build_vector_clock_ad_plan`) ET par le checker autonome
+  (`check.rs::verify_vector_clock_ad_plan`). En-têtes de modules
+  documentent la frontière.
+- Aucun test de rejet `UnadoptedStatefulRead` n'existait → ajout de deux
+  tests : `unadopted_stateful_read_rejected_on_producer_path` et
+  `unadopted_stateful_read_rejected_through_checker_entry_alone` (fixture :
+  domaine downsampling dont la valeur interne lit `delay1(input)` au rythme
+  audio au moment du tir — analogue corpus
+  `downsampling_02_domain_free_counter`). Le test checker assemble le plan
+  candidat à partir des dérivations directement, sans passer par la
+  vérification terminale du producteur.
+- Gates : transform 387 (385+2), vector_mode 35/35, clippy workspace
+  +1.97.0 vert, fmt vert. Arbitre en cours après commit.
